fix necessidade e comentario

+++ backend/src/main/java/com/utfpr/donare/service/ComentarioServiceImpl.java
@@ -1,5 +1,6 @@
 package com.utfpr.donare.service;
 
+import com.utfpr.donare.domain.User;
 import com.utfpr.donare.dto.ComentarioRequestDTO;
 import com.utfpr.donare.dto.ComentarioResponseDTO;
 import com.utfpr.donare.exception.ResourceNotFoundException;
@@ -7,10 +8,12 @@ import com.utfpr.donare.domain.Campanha;
 import com.utfpr.donare.domain.Comentario;
 import com.utfpr.donare.repository.CampanhaRepository;
 import com.utfpr.donare.repository.ComentarioRepository;
+import com.utfpr.donare.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,16 +22,28 @@ public class ComentarioServiceImpl implements ComentarioService {
 
     private final ComentarioRepository comentarioRepository;
     private final CampanhaRepository campanhaRepository;
+    private final UserRepository userRepository;
 
     @Override
-    public ComentarioResponseDTO criarComentario(Long idCampanha, ComentarioRequestDTO comentarioRequestDTO, String userEmail) {
+    public ComentarioResponseDTO criarComentario(Long idCampanha, ComentarioRequestDTO comentarioRequestDTO) {
         Campanha campanha = campanhaRepository.findById(idCampanha)
                 .orElseThrow(() -> new ResourceNotFoundException("Campanha não encontrada com o id: " + idCampanha));
 
+        User user = userRepository.findByEmail(comentarioRequestDTO.getUserEmail())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario não encontrada com o email: " + comentarioRequestDTO.getUserEmail()));
+
+
         Comentario comentario = new Comentario();
         comentario.setConteudo(comentarioRequestDTO.getConteudo());
-        comentario.setUserEmail(userEmail);
-        comentario.setIdComentarioPai(comentarioRequestDTO.getIdComentarioPai());
+        comentario.setCampanha(campanha);
+        comentario.setUser(user);
+
+        // Trata o comentário pai, se houver
+        if (comentarioRequestDTO.getIdComentarioPai() != null) {
+            Comentario comentarioPai = comentarioRepository.findById(comentarioRequestDTO.getIdComentarioPai())
+                    .orElseThrow(() -> new ResourceNotFoundException("Comentário pai não encontrado com o id: " + comentarioRequestDTO.getIdComentarioPai()));
+            comentario.setComentarioPai(comentarioPai);
+        }
 
         Comentario comentarioSalvo = comentarioRepository.save(comentario);
         return converterParaResponseDTO(comentarioSalvo);
@@ -55,22 +70,38 @@ public class ComentarioServiceImpl implements ComentarioService {
     }
 
     @Override
-    public ComentarioResponseDTO editarComentario(Long idComentario,ComentarioRequestDTO comentarioRequestDTO, String userEmail) {
+    public ComentarioResponseDTO editarComentario(Long idComentario,ComentarioRequestDTO comentarioRequestDTO) {
         Comentario comentario = comentarioRepository.findById(idComentario)
                 .orElseThrow(() -> new ResourceNotFoundException("Comentario não encontrada com o id: " + idComentario));
 
-        comentario.setConteudo(comentarioRequestDTO.getConteudo());
-        comentario.setUserEmail(userEmail);
+        User user = userRepository.findByEmail(comentarioRequestDTO.getUserEmail())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario não encontrada com o email: " + comentarioRequestDTO.getUserEmail()));
+
+        if(comentario.getUser().getEmail().equals(comentarioRequestDTO.getUserEmail())){
+            comentario.setConteudo(comentarioRequestDTO.getConteudo());
+            comentario.setUser(user);
+        }
+        else{
+            throw new ResourceNotFoundException("Usuário com e-mail inválido para editar esse comentário" + comentarioRequestDTO.getUserEmail());
+        }
 
         Comentario comentarioAtualizado = comentarioRepository.save(comentario);
         return converterParaResponseDTO(comentarioAtualizado);
     }
 
     @Override
-    public void deletarComentario(Long idComentario, String userEmail) {
+    public void deletarComentario(Long idComentario, ComentarioRequestDTO comentarioRequestDTO) {
         Comentario comentario = comentarioRepository.findById(idComentario)
-                .orElseThrow(() -> new ResourceNotFoundException("Postagem não encontrada com o id: " + idComentario));
-        comentarioRepository.delete(comentario);
+                .orElseThrow(() -> new ResourceNotFoundException("Comentário não encontrado com o id: " + idComentario));
+
+        if(comentario.getUser().getEmail().equals(comentarioRequestDTO.getUserEmail())){
+            comentarioRepository.delete(comentario);
+        }
+        else{
+            throw new ResourceNotFoundException("Usuário com e-mail inválido para excluir esse comentário" + comentarioRequestDTO.getUserEmail());
+        }
+
+
     }
 
     private ComentarioResponseDTO converterParaResponseDTO(Comentario comentario) {
@@ -78,14 +109,19 @@ public class ComentarioServiceImpl implements ComentarioService {
         String organizadorCampanha = (campanha != null && campanha.getOrganizador() != null)
                 ? campanha.getOrganizador() : "Organizador não disponível";
 
+        Long idComentarioPai = comentario.getComentarioPai() != null
+                ? comentario.getComentarioPai().getId()
+                : null;
+
+        assert comentario.getCampanha() != null;
         return new ComentarioResponseDTO(
                 comentario.getId(),
                 comentario.getConteudo(),
                 comentario.getDataCriacao(),
                 comentario.getUser().getId(),
-                comentario.getUserEmail(),
+                comentario.getUser().getEmail(),
                 comentario.getUser().getNome(),
-                comentario.getIdComentarioPai(),
+                idComentarioPai ,
                 comentario.getCampanha().getId());
     }
 }
